requirements.txt
asgiref==3.10.0
certifi==2025.11.12
charset-normalizer==3.4.4
Django==5.2.8
django-cors-headers==4.9.0
django-rest-framework==0.1.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
gunicorn==23.0.0
idna==3.11
mysqlclient==2.2.7
numpy==2.2.6
opencv-python==4.12.0.88
opencv-python-headless==4.12.0.88
packaging==25.0
pika==1.3.2
pillow==12.0.0
psycopg2-binary==2.9.11
PyJWT==2.10.1
pytesseract==0.3.13
python-dateutil==2.9.0.post0
RapidFuzz==3.14.3
requests==2.32.5
six==1.17.0
sqlparse==0.5.3
urllib3==2.5.0
